renamed domainUtils.jsx to urlDomainAnalysis.jsx
