+++ src/components/NavItems/Assistant/utils/urlDomainAnalysisUtils.jsx
